Fix Mixed Content violation with jQuery
The GitHub pages don't work right now with modern browsers because there is a Mixed Content violation. JavaScript files need to be loaded over HTTPS if the visiting page is also server with HTTPS, this should be resolved with this commit.
This commit also updates the jQuery library to the latest v1 version. All v1 versions should be compatible with each other.